pyproject.toml
[project]
name = "zlabel-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "numpy>=2.2.5",
    "onnxruntime-gpu>=1.22.0",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=11.2.1",
    "requests>=2.32.3",
    "rich>=14.0.0",
]
